refactor: isolate policy plugin render planning
